+++ src/cow-react/modules/wallet/api/components/WalletModal/options/CoinbaseWalletOption.tsx
@@ -1,7 +1,7 @@
 import { Connector } from '@web3-react/types'
 import COINBASE_ICON_URL from '../../../assets/coinbase.svg'
 import { coinbaseWalletConnection, ConnectionType } from 'connection'
-import { getConnectionName } from 'connection/utils'
+import { getConnectionName } from '@cow/modules/wallet/api/utils'
 
 import Option from '../Option'
 import { useSelectedWallet } from 'state/user/hooks'
